+++ Packages/Syntax/Package.swift
@@ -42,7 +42,7 @@ let package = Package(
         .package(url: "https://github.com/1024jp/tree-sitter-python", branch: "swiftPackage"),
         .package(url: "https://github.com/tree-sitter/tree-sitter-ruby", from: Version(0, 23, 1)),
         .package(url: "https://github.com/tree-sitter/tree-sitter-rust", from: Version(0, 24, 0)),
-        .package(url: "https://github.com/tree-sitter/tree-sitter-scala", from: Version(0, 24, 0)),
+        .package(url: "https://github.com/tree-sitter/tree-sitter-scala", from: Version(0, 26, 2)),
         .package(url: "https://github.com/DerekStride/tree-sitter-sql", branch: "gh-pages"),
         .package(url: "https://github.com/alex-pinkus/tree-sitter-swift", branch: "with-generated-files"),
         .package(url: "https://github.com/tree-sitter/tree-sitter-typescript", from: Version(0, 23, 2)),